requirements.txt
anyio==3.7.1
asgiref==3.4.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Django==4.0.1
exceptiongroup==1.1.3
Flask==2.3.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-telegram-bot==20.4
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.2
urllib3==2.0.4
uWSGI==2.0.22
Werkzeug==2.3.7
